Simplify blitting and font-related code.
Use characters as the index for char-based fonts in graphics.py. This allows most
code, which uses chars as the image of its entity, to simplify its representation of this
quite a bit.

Return handles from GetChar() (consistency).

Define GridBlit in the SurfaceRegistry (just so we don't have to pass the ID and reg).