Stop using the Properties enum.
The idea was that an enum would help catch typoes more than arbitrary strings
and that the language would handle these errors so I didn't have to. In reality,
Python enums are too cumbersome for this use case and I feel better off
writing my own code that detects when a component name is different
from that I registered to the entity class.

Using __getattr__ and __setattr__ for this... we'll see whether that's smart
or not.