Decide attack actions in main.py
The original idea was that an item would define an attack type and
action which would then feed into the path/action finding algorithm in a
modular way. The problem is that this doesn't scale to holding weapons
where N > 1 and it means the player can't attack if N = 0. There was an
idea to have the player equip weapons, but this is antithetical to the
game's core idea.

Instead, going forward, I will probably want weapons to at most change a
generic attack stat, modify the attack range (+1 or such), or change the
attack type with a preference for some types over others.

I also plan to simplify damage calculations, but haven't decided on how
yet.