Remove the "Modifier" abstraction. (1/2)
This ends up feeling messy, unstructured, and actually makes things a
bit harder to work with by many modifiers trying to describe game
functionality in code areas that shouldn't be concerned with that.

This is the first commit of two since to make this work, I had to
comment out several unused functionalities.