ActionManager: A better way to organize actions.
The current strategy just won't scale to more complex interactions such
as if an attack causes a reaction in the defender that needs to process
before the turn can end or the player initiates another action. This,
naturally, leads to the idea of ordered actions held in a stack, and
independent actions like visual animations which have no effect on
state.